refactor(server): carve claude-tui-session.js into PTY + form driver modules
Pure-move extraction of the two most empirically-tuned seams out of the
4435-LOC claude-tui-session.js (#5559), with zero behaviour change:

- packages/server/src/claude-tui/pty-driver.js — PTY-write / paste-throttle
  layer: ANSI_STRIP, formatHexDump, CLAUDE, the AUTH_* classifiers,
  ensureCwdTrusted, writeHookSettings, and the throttled-write methods
  (_writePtyTextThrottled, _writePtyMultiQuestionSequence,
  _writePtyArrowNavSequence) on PtyDriverMixin.
- packages/server/src/claude-tui/form-driver.js — interactive AskUserQuestion
  driver: the watchdog/settle constants and respondToQuestion +
  _onAskUserQuestionStall + _teardownAskUserQuestion on FormDriverMixin.

The mixin methods are copied onto ClaudeTuiSession.prototype via applyMixin()
(own-descriptor copy, preserving non-enumerability), so `this` and every
static reference resolve exactly as before. Method bodies and the pinned byte
sequences are moved byte-identically; the only changes are import/export
mechanics and the applyMixin wiring. writeHookSettings is re-exported from
claude-tui-session.js so all existing imports keep working unchanged.

The full claude-tui test suite (301 + 39) passes unmodified.

Closes #5559
